+++ src/validators/numeric/index.ts
@@ -2,3 +2,4 @@ export { IsInteger } from "./is-integer";
 export { IsFinite } from "./is-finite";
 
 export { IsPositive } from "./is-positive";
+export { IsNegative } from "./is-negative";